Exercise 4.21 complete

+++ part4/controllers/blogs.js
@@ -4,13 +4,7 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if(authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
@@ -19,7 +13,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if(!decodedToken.id) {
     return response.status(401).json({ error: 'invalid token' })
   }
@@ -49,8 +43,21 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
 
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(id)
+  if(blog.user.toString() !== user._id.toString()) {
+    return response.status(400).json({ error: 'invalid user' })
+  }
+
+  user.blogs = user.blogs.filter(b => b.toString() !== blog._id.toString())
+  await user.save()
   await Blog.findByIdAndDelete(id)
+  
   return response.status(204).end()
 })
